+++ router/core/router.go
@@ -101,7 +101,7 @@ type (
 	}
 
 	SubgraphTransportOptions struct {
-		TransportRequestOptions
+		*TransportRequestOptions
 		SubgraphMap map[string]*TransportRequestOptions
 	}
 
@@ -1664,19 +1664,36 @@ func DefaultFileUploadConfig() *config.FileUpload {
 	}
 }
 
-func NewTransportRequestOptions(cfg config.GlobalSubgraphRequestRule) TransportRequestOptions {
-	return TransportRequestOptions{
-		RequestTimeout:         cfg.RequestTimeout,
-		ResponseHeaderTimeout:  cfg.ResponseHeaderTimeout,
-		ExpectContinueTimeout:  cfg.ExpectContinueTimeout,
-		KeepAliveIdleTimeout:   cfg.KeepAliveIdleTimeout,
-		DialTimeout:            cfg.DialTimeout,
-		TLSHandshakeTimeout:    cfg.TLSHandshakeTimeout,
-		KeepAliveProbeInterval: cfg.KeepAliveProbeInterval,
+func NewTransportRequestOptions(cfg config.GlobalSubgraphRequestRule) *TransportRequestOptions {
+	defaults := DefaultTransportRequestOptions()
 
-		MaxConnsPerHost:     cfg.MaxConnsPerHost,
-		MaxIdleConns:        cfg.MaxIdleConns,
-		MaxIdleConnsPerHost: cfg.MaxIdleConnsPerHost,
+	return &TransportRequestOptions{
+		RequestTimeout:         or(cfg.RequestTimeout, defaults.RequestTimeout),
+		TLSHandshakeTimeout:    or(cfg.TLSHandshakeTimeout, defaults.TLSHandshakeTimeout),
+		ResponseHeaderTimeout:  or(cfg.ResponseHeaderTimeout, defaults.ResponseHeaderTimeout),
+		ExpectContinueTimeout:  or(cfg.ExpectContinueTimeout, defaults.ExpectContinueTimeout),
+		KeepAliveProbeInterval: or(cfg.KeepAliveProbeInterval, defaults.KeepAliveProbeInterval),
+		KeepAliveIdleTimeout:   or(cfg.KeepAliveIdleTimeout, defaults.KeepAliveIdleTimeout),
+		DialTimeout:            or(cfg.DialTimeout, defaults.DialTimeout),
+		MaxConnsPerHost:        or(cfg.MaxConnsPerHost, defaults.MaxConnsPerHost),
+		MaxIdleConns:           or(cfg.MaxIdleConns, defaults.MaxIdleConns),
+		MaxIdleConnsPerHost:    or(cfg.MaxIdleConnsPerHost, defaults.MaxIdleConnsPerHost),
+	}
+}
+
+func DefaultTransportRequestOptions() *TransportRequestOptions {
+	return &TransportRequestOptions{
+		RequestTimeout:         60 * time.Second,
+		TLSHandshakeTimeout:    10 * time.Second,
+		ResponseHeaderTimeout:  0 * time.Second,
+		ExpectContinueTimeout:  0 * time.Second,
+		KeepAliveProbeInterval: 30 * time.Second,
+		KeepAliveIdleTimeout:   0 * time.Second,
+		DialTimeout:            30 * time.Second,
+
+		MaxConnsPerHost:     100,
+		MaxIdleConns:        1024,
+		MaxIdleConnsPerHost: 20,
 	}
 }
 
@@ -1687,8 +1704,7 @@ func NewSubgraphTransportOptions(cfg config.TrafficShapingRules) *SubgraphTransp
 	}
 
 	for k, v := range cfg.Subgraphs {
-		opts := NewTransportRequestOptions(*v)
-		base.SubgraphMap[k] = &opts
+		base.SubgraphMap[k] = NewTransportRequestOptions(*v)
 	}
 
 	return base
@@ -1696,20 +1712,8 @@ func NewSubgraphTransportOptions(cfg config.TrafficShapingRules) *SubgraphTransp
 
 func DefaultSubgraphTransportOptions() *SubgraphTransportOptions {
 	return &SubgraphTransportOptions{
-		TransportRequestOptions: TransportRequestOptions{
-			RequestTimeout:         60 * time.Second,
-			TLSHandshakeTimeout:    10 * time.Second,
-			ResponseHeaderTimeout:  0 * time.Second,
-			ExpectContinueTimeout:  0 * time.Second,
-			KeepAliveProbeInterval: 30 * time.Second,
-			KeepAliveIdleTimeout:   0 * time.Second,
-			DialTimeout:            30 * time.Second,
-
-			MaxConnsPerHost:     100,
-			MaxIdleConns:        1024,
-			MaxIdleConnsPerHost: 20,
-		},
-		SubgraphMap: map[string]*TransportRequestOptions{},
+		TransportRequestOptions: DefaultTransportRequestOptions(),
+		SubgraphMap:             map[string]*TransportRequestOptions{},
 	}
 }
 
@@ -1814,6 +1818,7 @@ func WithApolloCompatibilityFlagsConfig(cfg config.ApolloCompatibilityFlags) Opt
 			cfg.SuppressFetchErrors.Enabled = true
 			cfg.ReplaceUndefinedOpFieldErrors.Enabled = true
 			cfg.ReplaceInvalidVarErrors.Enabled = true
+			cfg.ReplaceValidationErrorStatus.Enabled = true
 		}
 		r.apolloCompatibilityFlags = cfg
 	}
@@ -1839,7 +1844,7 @@ func WithCacheWarmupConfig(cfg *config.CacheWarmupConfiguration) Option {
 
 type ProxyFunc func(req *http.Request) (*url.URL, error)
 
-func newHTTPTransport(opts TransportRequestOptions, proxy ProxyFunc) *http.Transport {
+func newHTTPTransport(opts *TransportRequestOptions, proxy ProxyFunc) *http.Transport {
 	dialer := &net.Dialer{
 		Timeout:   opts.DialTimeout,
 		KeepAlive: opts.KeepAliveProbeInterval,
@@ -2012,4 +2017,11 @@ func MetricConfigFromTelemetry(cfg *config.Telemetry) *rmetric.Config {
 			ExcludeMetricLabels: cfg.Metrics.Prometheus.ExcludeMetricLabels,
 		},
 	}
+}
+
+func or[T any](maybe *T, or T) T {
+	if maybe != nil {
+		return *maybe
+	}
+	return or
 }